TabSurvey/requirements.txt
optuna==3.3.0
scikit-learn==1.0.2
pandas==1.3.5
configargparse==1.7
pytorch-tabnet==4.1.0
tensorflow-gpu==1.15.0
xgboost==1.6.2
catboost==1.2.2
lightgbm==4.1.0
einops==0.6.1
requests==2.31.0
qhoptim==1.1.0
yacs==0.1.8
lifelines==0.27.8
tabulate
natsort
https://github.com/ntruongn/model-trees/archive/master.zip
https://github.com/ntruongn/nam/archive/main.zip
ray[tune]